Remove obsolete SdkType
I keep confusing it with its replacement
`org.elixir_lang.sdk.elixir.Type`!
